Introduced `update(Byte)` and `done(): Byte`
This is very GC friedly changes, sometime we need insert just a single byte into
hash, and it is very strange idea to wrap it into single-element array.

And when we need to get a hash value as single byte (for checksum for example),
we can do it without creating a lot of internal arrays.